refactor(python): Move `PythonInspector` unrelated utilities out
Separate generic Python utilities from those specific to `PythonInspector`
by moving them to a dedicated file.
